Refatora e corrige inconsistencias no projeto (#54)
- Remove endpoints de endereco

- Sincroniza metodo HTTP e status code nos endpoints

- Remove `/delete` de endpoints de DELETE

- Corrige inconsistencias nos tipos de dado

- Renomeia classe principal da aplicacao

+++ src/main/java/com/recrutaibackend/dto/InstitutionResponse.java
@@ -9,7 +9,7 @@ public record InstitutionResponse(
         String headline,
         UserResponse owner,
         String industry,
-        int companySizeId,
+        short companySizeId,
         int associatedEmployees,
         AddressResponse headquarters,
         String website,
