+++ src/main/java/com/recrutaibackend/dto/SchoolResponse.java
@@ -3,7 +3,7 @@ package com.recrutaibackend.dto;
 public record SchoolResponse(
         long id,
         InstitutionResponse institution,
-        int schoolSizeId,
+        short schoolSizeId,
         int associatedAlumni
 ) {
 }
